Add files via upload
Note: 
Before building the docker container, our database needs to be connected and a new password should be set for security purposes.
Once this is done, you can build the docker instance and run it. Once the Postgres DB is connected, you can pinpoint the container to which exact column content you want to have hashed.